Working on adding 64 bit

+++ Defines/FMProcess.cs
@@ -1,20 +1,18 @@
 ﻿using System;
-using System.Collections.Generic;
-using System.Text;
 using System.Diagnostics;
 
 namespace FMScoutFramework.Core
 {
-	/// <summary>
-	/// Holder for the Football Manager process
-	/// </summary>
-	public class FMProcess
-	{
-		public Process Process { get; set; }
-		public IntPtr Pointer { get; set; }
-		public int EndPoint { get; set; }
-		public int BaseAddress { get; set; }
-		public string VersionDescription { get; set; }
-		public int HeapAddress { get; set; }
-	}
+    /// <summary>
+    /// Holder for the Football Manager process
+    /// </summary>
+    public class FMProcess
+    {
+        public Process Process { get; set; }
+        public IntPtr Pointer { get; set; }
+        public int EndPoint { get; set; }
+        public IntPtr BaseAddress { get; set; }
+        public string VersionDescription { get; set; }
+        public int HeapAddress { get; set; }
+    }
 }
